+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 - Print
+- Beautify the interface
+- Show characters & line numbers on status bar
 
 ## 0.0.1-rc.1